+++ storage/storage_test.go
@@ -9,54 +9,66 @@ import (
 	"github.com/philippgille/ln-paywall/wall"
 )
 
-// testStorageClient tests if reading and writing to the storage works properly.
+type foo struct {
+	Bar string
+}
+
+// testStorageClient tests if reading from and writing to the storage works properly.
 func testStorageClient(storageClient wall.StorageClient, t *testing.T) {
 	key := strconv.FormatInt(rand.Int63(), 10)
 
-	// Initially the key shouldn't exist, thus "was used" should be false
-	expected := false
-	actual, err := storageClient.WasUsed(key)
+	// Initially the key shouldn't exist
+	found, err := storageClient.Get(key, new(foo))
 	if err != nil {
 		t.Error(err)
 	}
-	if actual != expected {
-		t.Errorf("Expected: %v, but was: %v", expected, actual)
+	if found {
+		t.Errorf("A value was found, but no value was expected")
 	}
 
-	// Set the "key" to be used
-	err = storageClient.SetUsed(key)
+	// Store an object
+	val := foo{
+		Bar: "baz",
+	}
+	err = storageClient.Set(key, val)
 	if err != nil {
 		t.Error(err)
 	}
 
-	// Check usage again, this time "was used" should be true
-	expected = true
-	actual, err = storageClient.WasUsed(key)
+	// Retrieve the object
+	expected := val
+	actualPtr := new(foo)
+	found, err = storageClient.Get(key, actualPtr)
 	if err != nil {
 		t.Error(err)
 	}
+	if !found {
+		t.Errorf("No value was found, but should have been")
+	}
+	actual := *actualPtr
 	if actual != expected {
 		t.Errorf("Expected: %v, but was: %v", expected, actual)
 	}
 }
 
 // interactWithStorage reads from and writes to the DB. Meant to be executed in a goroutine.
-// Does NOT check if the DB works correctly (that's done by another test), only checks for errors
+// Does NOT check if the DB works correctly (that's done elsewhere),
+// only checks for errors that might occur due to concurrent access.
 func interactWithStorage(storageClient wall.StorageClient, key string, t *testing.T, waitGroup *sync.WaitGroup) {
 	defer waitGroup.Done()
 
 	// Read
-	_, err := storageClient.WasUsed(key)
+	_, err := storageClient.Get(key, new(foo))
 	if err != nil {
 		t.Error(err)
 	}
 	// Write
-	err = storageClient.SetUsed(key)
+	err = storageClient.Set(key, foo{})
 	if err != nil {
 		t.Error(err)
 	}
 	// Read
-	_, err = storageClient.WasUsed(key)
+	_, err = storageClient.Get(key, new(foo))
 	if err != nil {
 		t.Error(err)
 	}