+++ wall/echo.go
@@ -1,6 +1,7 @@
 package wall
 
 import (
+	"encoding/hex"
 	"fmt"
 	"log"
 	"net/http"
@@ -36,18 +37,27 @@ func NewEchoMiddleware(invoiceOptions InvoiceOptions, lnClient LNclient, storage
 						Internal: err,
 					}
 				}
+
+				// Cache the invoice metadata
+				invoiceID := hex.EncodeToString([]byte(invoice.PaymentHash))
+				metadata := invoiceMetaData{
+					Method: ctx.Request().Method,
+					Path:   ctx.Request().URL.Path,
+				}
+				storageClient.Set(invoiceID, metadata)
+
 				stdOutLogger.Printf("Sending invoice in response: %v", invoice)
 				ctx.Response().Header().Set("Content-Type", "application/vnd.lightning.bolt11")
 				ctx.Response().Status = http.StatusPaymentRequired
 				// The actual invoice goes into the body
-				ctx.Response().Write([]byte(invoice))
+				ctx.Response().Write([]byte(invoice.PaymentRequest))
 				return &echo.HTTPError{
 					Code:    http.StatusPaymentRequired,
 					Message: invoice,
 				}
 			}
 			// Check if the provided preimage belongs to a settled API payment invoice and that it wasn't already used. Also store used preimages.
-			invalidPreimageMsg, err := handlePreimage(preimageHex, storageClient, lnClient)
+			invalidPreimageMsg, err := handlePreimage(ctx.Request(), storageClient, lnClient)
 			if err != nil {
 				errorMsg := fmt.Sprintf("An error occurred during checking the preimage: %+v", err)
 				log.Printf("%v\n", errorMsg)